Introduce NodeInfo.informeesOfNode
Currently the code to compute the informees of a transaction node is
duplicated to several places. This adds a generic function to compute
this that works for different representations of transaction nodes
and allows reusing one common function for computing the informees
and from it, the transaction witnesses.

While writing this I noticed that the scenario ledger did not consider
the submitter an initial witness and hence we had e.g. fetch nodes
without any witnesses. This code fixes this and includes the fixes
to the tests. The main impact of this bug was missing "known to" line
in scenario results for top-level fetch nodes, which are only creatable
with scenarios.

The informees of a LookupByKey are still subtly wrong: the signatories
of the fetched contract should be informed, but we only inform the key
maintainers. Fix to that will be in another PR as it requires bigger changes
to transaction.

CHANGELOG_BEGIN
- [DAML SDK] Fix computation of witnesses of top-level fetch nodes in scenario results ("known to").
CHANGELOG_END