standard change extract script
As part of our SOC2/ISO certification, we need to be able to evidence a
list of "Standard Changes" for the DAML SDK project. This commit adds a
script that extracts; for a given month, the list of Standard Changes
that happened along with relevant information (author, reviewer, etc.).

CHANGELOG_BEGIN
CHANGELOG_END